Refactoring wip
