+++ nomos-blend/message/src/crypto/proofs/quota/mod.rs
@@ -1,16 +1,27 @@
+use core::array::from_ref;
+
+use groth16::Fr;
 use nomos_core::crypto::ZkHash;
+use num_bigint::BigUint;
 use serde::{Deserialize, Serialize};
 
-use crate::crypto::proofs::quota::inputs::{PrivateInputs, PublicInputs};
+use crate::crypto::{
+    proofs::quota::inputs::{PrivateInputs, PublicInputs},
+    pseudo_random_sized_bytes,
+};
 
 pub mod inputs;
+
+const KEY_NULLIFIER_SIZE: usize = size_of::<ZkHash>();
+const PROOF_CIRCUIT_SIZE: usize = 128;
+pub const PROOF_OF_QUOTA_SIZE: usize = KEY_NULLIFIER_SIZE.checked_add(PROOF_CIRCUIT_SIZE).unwrap();
 
 #[derive(Debug, Clone, Copy, Hash, Serialize, Deserialize, PartialEq, Eq)]
 pub struct ProofOfQuota {
     #[serde(with = "groth16::serde::serde_fr")]
     key_nullifier: ZkHash,
     #[serde(with = "serde_big_array::BigArray")]
-    proof: [u8; 128],
+    proof: [u8; PROOF_CIRCUIT_SIZE],
 }
 
 impl ProofOfQuota {
@@ -21,9 +32,13 @@ impl ProofOfQuota {
     }
 
     #[must_use]
-    pub fn verify(self, _public_inputs: &PublicInputs) -> bool {
+    pub fn verify(self, _public_inputs: &PublicInputs) -> Result<ZkHash, ()> {
         // TODO: Interact with circom circuit verification.
-        self == Self::dummy()
+        if self == Self::dummy() {
+            Ok(self.key_nullifier)
+        } else {
+            Err(())
+        }
     }
 
     // TODO: Remove this once the actual proof of quota is implemented.
@@ -32,6 +47,15 @@ impl ProofOfQuota {
         Self {
             key_nullifier: ZkHash::default(),
             proof: [0u8; _],
+        }
+    }
+}
+
+impl From<[u8; PROOF_OF_QUOTA_SIZE]> for ProofOfQuota {
+    fn from(value: [u8; PROOF_OF_QUOTA_SIZE]) -> Self {
+        Self {
+            key_nullifier: BigUint::from_bytes_be(&value[..KEY_NULLIFIER_SIZE]).into(),
+            proof: <[u8; PROOF_CIRCUIT_SIZE]>::try_from(&value[KEY_NULLIFIER_SIZE..]).unwrap(),
         }
     }
 }
