+++ nomos-blend/message/src/lib.rs
@@ -6,4 +6,4 @@ mod message;
 
 pub use encap::MessageIdentifier;
 pub use error::Error;
-pub use message::PayloadType;
+pub use message::payload::PayloadType;
