+++ nomos-blend/message/src/message/mod.rs
@@ -1,0 +1,6 @@
+pub mod blending_header;
+pub use self::blending_header::BlendingHeader;
+pub mod payload;
+pub use self::payload::Payload;
+pub mod public_header;
+pub use self::public_header::PublicHeader;
